Move Executable class to Executable.scala.
